Moved UIApplication+topViewController extension into dedicated source code file for improved organization.
